Refactor : 파일명 수정 및 title props 삭제
