+++ src/components/ImageCarouselArea/styles.ts
